WNM: allow specifying dialog-token
This commit adds the ability to specify the dialog token of a WNM BSS
Transition request frame via the hostapd control socket.

For this, the new 'dialog_token' option can be used. It accepts values
as a 8 bit unsigned integer. If not specified, the dialog token is set
to 1 like before.

Signed-off-by: David Bauer <mail@david-bauer.net>