ctrl: enable roaming between OWE APs
This allows to use the roam command with wpa_cli to force roaming on a
transition network. Previously, this was not possible, as the open SSID
is stored for the connection profile. Add a new function to also return
OWE transition networks if the profile SSID is set as the
transition-ssid for the OWE RSN network.

Signed-off-by: David Bauer <mail@david-bauer.net>