scan: stored last-seen transition SSID
Store the last seen transition SSID when conneted to an OWE network
enabling transition-mode.

This speeds up discovery of hidden transition networks when connected on
6GHz. As 6 GHz does not use inter-band transition mode operation,
the transition SSID name is missing from beacons on this band.

When the scan-list times out, wpa_supplicant has to scan two consecutive
times to first discover transition networks on 2.4 and 5GHz.

Optimize this by storing the last transition-SSID independent from the
scan-results.

Signed-off-by: David Bauer <mail@david-bauer.net>