bss: enable discovery of SSID for known OWE BSS
OWE BSSIDs are beaconed with a hidden SSID set, including them with a
hidden SSID in the scan_results when not explicitly proed for.

Currently wpa_supplicant creates suplicate entries for these BSSIDs in
the scan-list instead of updating the hidden results with the learnt
SSID when probing on a connection for OWE networks.

Update existing entries with hidden SSID instead. Also update entries
with learnt SSID when receiving a beacon frame for this SSID.

This is required to consistently fix roaming when connected to an OWE
network.

Signed-off-by: David Bauer <mail@david-bauer.net>